% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest_genome.R
\name{digest_genome}
\alias{digest_genome}
\title{Performs in silico digestion of genomes}
\arguments{
\item{genome}{String indicating the species to be loaded, e.g. 'human' will
load the package 'BSgenome.Hsapiens.UCSC.hg38'}

\item{enzyme}{String indicating the enzyme to be used to detect sites}
}
\value{
Returns data frame object with chromosome name, start position,
end position, and length of fragment for each gap between restriction digest sites
}
\description{
This function performs an in silico digestion of BSgenomes based on chosen
type II restriction endonucleases
}
