% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome.R
\name{get_genome}
\alias{get_genome}
\title{Choose a genome build}
\usage{
get_genome(genome)
}
\arguments{
\item{genome}{String indicating the species to be loaded, e.g. 'human' will
load the package 'BSgenome.Hsapiens.UCSC.hg38'}
}
\value{
Allows main digest function to use correct BSgenome package
}
\description{
This function adds the appropriate BSgenome package based on the users input
}
